Restore query module.
